feat: use uint32 for indices

+++ include/rmilib/raw_mesh.hpp
@@ -44,4 +44,4 @@ private:
 
 typedef RawMesh<double, size_t> TriangularMesh;
 
-typedef RawMesh<float, unsigned short> WebGLMesh;
+typedef RawMesh<float, unsigned int> WebGLMesh;
