+++ src/stl_reader.cpp
@@ -98,4 +98,4 @@ RawMesh<float_t, index_t> read_raw_triangular_mesh_stl(std::istream& stream) {
 }
 
 template TriangularMesh read_raw_triangular_mesh_stl<double, size_t>(std::istream& stream);
-template WebGLMesh read_raw_triangular_mesh_stl<float, unsigned short>(std::istream& stream);
+template WebGLMesh read_raw_triangular_mesh_stl<float, unsigned int>(std::istream& stream);
